refactoring unit test structure
